Rewrote the world model.
Up til now the world objet was a "fake" object. it was mostly a convenient way to store a bunch of methods that allowed the user to access some stats about the world. This is no longer the case.

The World model now is a proper database table. It has a unique ID we can use for permalinks and keeps some stats internally. This change required changing every single piece of code that dealt with worlds to fetch the object from the db instead of creating a new world object.