Intiial work on the make-checkins page.
From this page we should be able to create arbitrary checkins to test edge conditions. This prompted a rewrite (have I said that a lot?) of the API code since it was broken and had no notion of security. We are about 50% of the way to completion. The UI is nearly done. The main thing missing is the code that actually builds the payload and submits it. Validation is about 80% done. A couple of fields are missing but the core is in working order.