Reworked the logging system.
With the improved logging system we want to be able to send logs from the computers to the checkin server. At the moment we have untested code that should allow us to do just that.

After making sure that it works (it probably doesn't) we need to add the serverside support for that.